Move spec file for front
