+++ front/__tests__/reducers/Community.spec.js
@@ -1,6 +1,6 @@
-import CommunityReducer from '../Community'
-import { addCommunity } from '../../actions/Community'
-import CommunityModel   from '../../models/Community'
+import CommunityReducer from '../../client/reducers/Community'
+import { addCommunity } from '../../client/actions/Community'
+import CommunityModel   from '../../client/models/Community'
 
 const model = (params) => {
   return new CommunityModel(params)
